backend/requirements.txt
fastapi==0.115.0
uvicorn==0.24.0
python-dotenv==1.0.0
google-genai==1.31.0
pydantic==2.5.0
python-multipart==0.0.6
